Start marking passages from Hugh of St Victor.
Currently using line numbers from Baron’s edition of ‘De quinque
septenis’, as the only standard references he provides are to the PL
and his base manuscript, but I don’t think this makes sense in the long
run: it looks like the text is going to be re-edited anyway, and it
seems very unlikely that they are going to include his line numbers.
Instead, therefore, I’m going to make a transcription of the base
manuscript and refer to the line numbers for this. Still not very
useful, but at least it’s has some historical basis.